CBG-1150 Address questions/suggestions about the REST-side changes.
