+++ rest/multipart.go
@@ -37,7 +37,11 @@ func ReadJSONFromMIME(headers http.Header, input io.ReadCloser, into interface{}
 	}
 
 	// Decode the body bytes into target structure.
-	err = base.JSONDecoderWithOptions(input, true, true).Decode(&into)
+	decoder := base.JSONDecoder(input)
+	decoder.DisallowUnknownFields()
+	decoder.UseNumber()
+	err = decoder.Decode(into)
+
 	if err != nil {
 		err = base.WrapJSONUnknownFieldErr(err)
 		if errors.Cause(err) == base.ErrUnknownField {
@@ -47,44 +51,6 @@ func ReadJSONFromMIME(headers http.Header, input io.ReadCloser, into interface{}
 		}
 	}
 	_ = input.Close()
-	return err
-}
-
-// ReadSanitizeConfigJSON parses a JSON body, unmarshals it into "DbConfig".
-// Closes the input io.ReadCloser once done. It expands environment variables and sanitizes the
-// input bytes by converting the back quotes into double-quotes, escaping literal backslashes,
-// newlines or double-quotes with backslashes.
-func ReadSanitizeConfigJSON(headers http.Header, input io.ReadCloser, config *DbConfig) error {
-	// Performs the Content-Type validation and Content-Encoding check.
-	input, err := processContentEncoding(headers, input)
-	if err != nil {
-		return err
-	}
-
-	// Read body bytes to sanitize the content and substitute environment variables.
-	content, err := ioutil.ReadAll(input)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = input.Close() }()
-
-	// Expand environment variables.
-	content = expandEnv(content)
-
-	// Convert the back quotes into double-quotes, escapes literal
-	// backslashes, newlines or double-quotes with backslashes.
-	content = base.ConvertBackQuotedStrings(content)
-
-	// Decode the body bytes into target structure.
-	err = base.JSONDecoderWithOptions(bytes.NewReader(content), true, true).Decode(config)
-	if err != nil {
-		err = base.WrapJSONUnknownFieldErr(err)
-		if errors.Cause(err) == base.ErrUnknownField {
-			err = base.HTTPErrorf(http.StatusBadRequest, "JSON Unknown Field: %s", err.Error())
-		} else {
-			err = base.HTTPErrorf(http.StatusBadRequest, "Bad JSON: %s", err.Error())
-		}
-	}
 	return err
 }
